+++ crates/client/starknet-block-import/src/validation.rs
@@ -1,0 +1,51 @@
+use std::sync::Arc;
+
+use blockifier::execution::contract_class::ContractClass as BlockifierCasmClass;
+use mc_db::sierra_classes_db::SierraClassesDb;
+use mp_transactions::DeclareTransactionV2;
+use sp_consensus::Error as ConsensusError;
+
+use crate::compilation::{
+    blockifier_casm_class_to_compiled_class_hash, blockifier_sierra_class_to_compiled_class_hash,
+};
+
+pub fn validate_declare_v2_transaction(
+    declare_v2: DeclareTransactionV2,
+    casm_class: BlockifierCasmClass,
+    sierra_classes_db: Arc<SierraClassesDb>,
+) -> Result<(), ConsensusError> {
+    // 1. Check if we have the according Sierra class stored locally (should have been saved during the
+    //    add declare rpc call)
+    let sierra_class = sierra_classes_db
+        .get_sierra_class(declare_v2.class_hash.into())
+        .map_err(|e| ConsensusError::Other(Box::new(e)))?
+        .ok_or_else(|| {
+            ConsensusError::StateUnavailable(format!(
+                "Could not find Sierra class locally {:?}",
+                declare_v2.class_hash.0
+            ))
+        })?;
+    let expected_casm_class_hash = declare_v2.compiled_class_hash.0;
+
+    // 2. Check if Casm class matches the compiled class hash in the transaction
+    let compiled_casm_class_hash =
+        blockifier_casm_class_to_compiled_class_hash(casm_class).map_err(|e| ConsensusError::Other(Box::new(e)))?;
+    if compiled_casm_class_hash != expected_casm_class_hash {
+        return Err(ConsensusError::ClientImport(format!(
+            "Mismatched class hash (compiling casm class from extrinsic): expected (in tx) {0:x}, got {1:x}",
+            expected_casm_class_hash, compiled_casm_class_hash
+        )));
+    }
+
+    // 3. Compile Sierra class to Casm class and check class hashes again
+    let compiled_sierra_class_hash =
+        blockifier_sierra_class_to_compiled_class_hash(sierra_class).map_err(|e| ConsensusError::Other(Box::new(e)))?;
+    if compiled_sierra_class_hash != expected_casm_class_hash {
+        return Err(ConsensusError::ClientImport(format!(
+            "Mismatched class hash (compiling sierra class from local db): expected (in tx) {0:x}, got {1:x}",
+            expected_casm_class_hash, compiled_sierra_class_hash
+        )));
+    }
+
+    Ok(())
+}